refactor: make repackDashboardPages be module
